requirements-dev.txt
pytest==3.0.5
pytest-mock==1.5.0
pytest-pep8==1.0.6
pytest-pythonpath==0.7.1
